changes while testing

+++ GitaWebProject/Models/DeletedProductModel.cs
@@ -8,8 +8,8 @@
         public bool MakeFlag { get; set; }
         public bool FinishedGoodsFlag { get; set; }
         public string? Color { get; set; }
-        public int SafetyStockLevel { get; set; }
-        public int ReorderPoint { get; set; }
+        public short SafetyStockLevel { get; set; }
+        public short ReorderPoint { get; set; }
         public decimal StandardCost { get; set; }
         public decimal ListPrice { get; set; }
         public string? Size { get; set; }
